% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset_advanced.R
\name{generate_dataset_advanced}
\alias{feat.complex}
\alias{feat.square}
\alias{feat.square.neg}
\alias{feat.sum.1.5}
\alias{feat.sum.6.10}
\alias{generate_dataset_advanced}
\alias{neg.feat.sum.6.10}
\title{Functions for generating simulated data}
\usage{
feat.sum.1.5(X)

feat.sum.6.10(X)

neg.feat.sum.6.10(X)

feat.complex(X)

feat.square(X)

feat.square.neg(X)

generate_dataset_advanced(with_feat = F, data_size = 100,
  ncol_gene_mat = 100, feat_m = NA, feat_d = NA, mu_cen = 1.3,
  gamma1 = c(0, 0), gamma2 = c(0, 0), lam_m = 1/15, lam_d = 1/20,
  norm_vcov = c(1, 0.5, 0.5, 1), cov = 0)
}
\arguments{
\item{X}{a matrix}

\item{X}{a matrix}

\item{X}{a matrix}

\item{X}{a matrix}

\item{X}{a matrix}

\item{X}{a matrix}
}
\description{
Functions for generating simulated data
}
\keyword{internal}

