% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernels}
\alias{GaussianKernel}
\alias{LinearKernel}
\alias{PolyKernel}
\alias{kernels}
\title{Kernel functions for kernscr}
\usage{
GaussianKernel(x, y, sigma = 1)

LinearKernel(x, y, rho = 0)

PolyKernel(x, y, a = 0, d = 2)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{sigma}{variance parameter for the gaussain kernel.}

\item{rho}{TODO correlation parameter for the linear kernel.
Default is 0.}

\item{a}{TODO of the polynomial for the \code{PolyKernel}. Default is \code{0}.}

\item{d}{degree of the polynomial for the \code{PolyKernel}. Default is \code{2} (Quadratic kernel)}
}
\description{
Compute kernel TODO
}
\keyword{internal}

