% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernscr-package.R
\docType{package}
\name{kernscr-package}
\alias{kernscr}
\alias{kernscr-package}
\title{kernscr: a package to perform Kernel Machine Score Test for Pathway Analysis in the Presence of Semi-Competing Risks}
\description{
Kernel Machine Score Test for Pathway Analysis in the Presence of Semi-Competing Risks
}
\details{
The main functions of the kernscr package is \code{\link{compute_all_tests}}
}
\references{
Neykov M, Hejblum BP, Sinnot JA,  Kernel Machine Score Test for
Pathway Analysis in the Presence of Semi-Competing Risks, \emph{submitted}, 2016.
}

