% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{lambda}
\alias{lambda_pert}
\title{Lambda and its perturbation}
\usage{
lambda(t, all_times, failures, gamma_vec, U)

lambda_pert(t, perturb_mat, all_times, failures, gamma_vec, U)
}
\description{
Lambda and its perturbation
}
\keyword{internal}

