% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_vec.R
\name{M_vec}
\alias{M_vec}
\alias{M_vec_pert}
\title{M vector and its perturbation}
\usage{
M_vec(t, all_times, failures, gamma_vec, U)

M_vec_pert(perturb_mat, t, all_times, failures, gamma_vec, U)
}
\description{
M vector and its perturbation

M vector perturbation
}
\keyword{internal}
