% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kernscr.R
\name{plot_kernscr_methodsplit}
\alias{plot_kernscr_kernelsplit}
\alias{plot_kernscr_methodsplit}
\title{Plotting functions used in the manuscript}
\usage{
plot_kernscr_methodsplit(raw_melted, adj_melted, kernel, method,
  pathway_names = TRUE, title = NULL, lower_threshold = -4)

plot_kernscr_kernelsplit(raw_melted, adj_melted, kernel, method,
  pathway_names = TRUE, title = NULL, lower_threshold = -4)
}
\description{
Plotting functions used in the manuscript
}
\keyword{internal}

