% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_stat_both.R
\name{sup_stat_both}
\alias{sup_stat_both}
\title{sup_stat_both}
\usage{
sup_stat_both(num_perts, Ws = NULL, data, set_U, Cov_e_M_e_D, rho = 1:40,
  l0 = NA, kernel = c("gaussian", "poly"), est_gamma = FALSE,
  pca_thres = NULL, ...)
}
\arguments{
\item{num_perts}{number of perturbations used}

\item{kernel}{a character string indicating which kernel should be used. Currently implemented are
\code{"linear"}, \code{"gaussian"} or \code{"poly"}.}

\item{est_gamma}{logical flag indicating whether \code{gamma} should be estimated. Default is \code{FALSE},
in which case \code{0} value is used for gamma.}

\item{pca_thres}{the threshold to be used for PCA. Default is \code{NULL}, in which case no PCA is performed.}

\item{...}{extra parameters specific to the type of kernel used}
}
\value{
a \code{data.frame}
}
\description{
sup_stat_both
}

