% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VTM.R
\name{VTM}
\alias{VTM}
\title{VTM}
\usage{
VTM(vc, dm)
}
\arguments{
\item{vc}{vector to be repeated}

\item{dm}{number of times that \code{vc} should be repeated}
}
\value{
a matrix of \code{dm} rows
}
\description{
Repeats a vector as identical rows of a matrix
}
\keyword{internal}
