% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerneval.R
\name{kernelEval}
\alias{gaussKernelEval}
\alias{gaussKernelEval.multipleSigmas}
\alias{genericKernelEval}
\alias{kernelEval}
\alias{linKernelEval}
\alias{polyKernelEval}
\alias{polyKernelEval.multiple}
\title{Evaluation of kernels}
\usage{
kernelEval(Z, K = c("linear", "poly", "gaussian"), ...)

linKernelEval(Z)

gaussKernelEval(Z, sigma = 1)

gaussKernelEval.multipleSigmas(Z, sigma = 1)

polyKernelEval(Z, a = 0, d = 1)

polyKernelEval.multiple(Z, a = 0, d = 1)

genericKernelEval(Z, K, ...)
}
\arguments{
\item{Z}{\code{p x n} size Gram matrix.}

\item{K}{which kernel is evaluated by \code{kerneval}.Possible values include currently implemented kernels
designated by a character string \code{"linear"}, \code{"poly"} and \code{"gaussian"}. Otherwise can be a user defined function.}

\item{...}{other arguments to be passed to the kenrel evaluated.}

\item{sigma}{sigma variance parameter for the \code{"gaussian"} kernel.}

\item{a}{TODO of the polynomial for the \code{"poly"}. Default is \code{0}.}

\item{d}{degree of the polynomial for the \code{"poly"}. Default is \code{2} (Quadratic kernel)}
}
\description{
Evaluation of kernels
}
\details{
\code{kernelEval} kernel eval works only for gaussian, polynomial and linear kernels currently.
To be implemented for other kernels too...

\code{genericKernelEval}
}
\keyword{internal}

