% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernscr-package.R
\docType{package}
\name{kernscr-package}
\alias{kernscr-package}
\alias{kernscr}
\title{kernscr: a package to perform Kernel Machine Score Test for Pathway Analysis in the Presence of Semi-Competing Risks}
\description{
Kernel Machine Score Test for Pathway Analysis in the Presence of Semi-Competing Risks
}
\details{
\tabular{ll}{
Package: \tab kernscr\cr
Type: \tab Package\cr
Version: \tab 1.0.6\cr
Date: \tab 2023-04-17\cr
License:\tab \href{http://www.gnu.org/licenses/gpl-2.0.txt}{GPL-2}\cr
}
The main function of the kernscr package is \code{\link{compute_all_tests}}
}
\references{
Neykov M, Hejblum BP, Sinnot JA, Kernel Machine Score Test for
Pathway Analysis in the Presence of Semi-Competing Risks, \emph{Stat Methods in Med Res},
, 27(4): 1099-1114 (2018). <doi: 10.1177/0962280216653427>.
}
\author{
Matey Neykov, Boris P. Hejblum, Jennifer A. Sinnott
--- Maintainer: Boris P. Hejblum
}
