% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernscr-package.R
\docType{package}
\name{kernscr-package}
\alias{kernscr}
\alias{kernscr-package}
\title{kernscr: a package to perform Kernel Machine Score Test for Pathway Analysis in the Presence of Semi-Competing Risks}
\description{
Kernel Machine Score Test for Pathway Analysis in the Presence of Semi-Competing Risks
}
\details{
\tabular{ll}{
Package: \tab kernscr\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2016-05-03\cr
License:\tab \href{http://www.gnu.org/licenses/gpl-2.0.txt}{GPL-2}\cr
}
The main function of the kernscr package is \code{\link{compute_all_tests}}
}
\author{
Boris P. Hejblum, Matey Neykov, Jennifer A. Sinnott
--- Maintainer: Boris P. Hejblum
}
\references{
Neykov M, Hejblum BP, Sinnot JA,  Kernel Machine Score Test for
Pathway Analysis in the Presence of Semi-Competing Risks, \emph{submitted}, 2016.
}

